add a microsecond sleep after exec of 'before-sleep'
when using before-sleep swayidle inhibits logind to go to sleep, before
it runs a command, typically swaylock.

but without this artificial microsecond sleep swayidle would release the
inhibit lock too fast, and it would happen (more often than not) for swaylock
to not have the time to spawn and lock the screen. then, on resume for a moment
the screen contents are visible before swaylock finally locks it.

the time is arbitrary selected, seemed long enough for computers and
short enough for people. seemed to work fine on my i7-5600U / MZ7TE256
sata ssd.

a more proper solution would be if swaylock would inform swayidle that
it did lock the screen, but we don't know that we run swaylock always,
and there's no generic protocol to know if a program is activated
(apart from systemd's one).