Drop logind integration
This integration is unwanted, as expressed in #117.

Additionally, the integration is not quite correct; to properly launch a
screen locker before sleeping, swayidle should set a suspend inhibitor
**at startup**, and upon the PrepareForSleep signal, it should start the
screen locker and then release the inhibitor lock. The existing
implementation is racey and may run the locker _after_ resuming from
suspend.

Fixes: https://github.com/swaywm/swayidle/issues/117
Closes: https://github.com/swaywm/swayidle/issues/127
Closes: https://github.com/swaywm/swayidle/issues/121
Closes: https://github.com/swaywm/swayidle/issues/64
Closes: https://github.com/swaywm/swayidle/issues/59
Closes: https://github.com/swaywm/swayidle/pull/125
Closes: https://github.com/swaywm/swayidle/pull/89