Avoid using wl_display_dispatch
wl_display_dispatch is a convenience function intended to poll the
socket and wait for events to be dispatched, and may block indefinitely.
This function is primarily intended for simple clients that do not use
event loops.

swayidle uses wl_event_loop to monitor the display socket, and only
calls the event handler if there is something to read. Instead of
polling again and possibly blocking, read whatever is available
directly.

Fixes: https://github.com/swaywm/swayidle/issues/193